fix(debug_playbook): breaks with ssh proxy
ansible-playbook seems to require *some* output or it thinks the SSH connection is broken:

```
SSH Error: data could not be sent to the remote host. Make sure this host can be reached over ssh
```

`cd` will return a small response but essentially do nothing.

Strangely, exit works fine with ansible itself and when not using an ssh tunnel/proxy. :/